update: fix config char type

+++ app/Http/Controllers/User/CommController.php
@@ -20,7 +20,7 @@ class CommController extends Controller
                 'withdraw_close' => (int)config('v2board.withdraw_close_enable', 0),
                 'currency' => config('v2board.currency', 'CNY'),
                 'currency_symbol' => config('v2board.currency_symbol', '¥'),
-                'commission_distribution_enable' => config('v2board.commission_distribution_enable', 0),
+                'commission_distribution_enable' => (int)config('v2board.commission_distribution_enable', 0),
                 'commission_distribution_l1' => config('v2board.commission_distribution_l1'),
                 'commission_distribution_l2' => config('v2board.commission_distribution_l2'),
                 'commission_distribution_l3' => config('v2board.commission_distribution_l3')
